Hotlink description.
